gov-monitoring/mobility-rfp-monitor-web/pyproject.toml
[project]
name = "mobility-rfp-monitor-web"
version = "0.1.0"
description = "Web service for monitoring Korean government mobility RFPs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
